+++ CoreVisualGraph/test/unit/src/au/gov/asd/tac/constellation/graph/visual/plugins/blaze/BlazeContextMenuNGTest.java
@@ -1,0 +1,113 @@
+/*
+ * Copyright 2010-2024 Australian Signals Directorate
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package au.gov.asd.tac.constellation.graph.visual.plugins.blaze;
+
+import au.gov.asd.tac.constellation.graph.GraphElementType;
+import au.gov.asd.tac.constellation.graph.GraphReadMethods;
+import au.gov.asd.tac.constellation.graph.StoreGraph;
+import au.gov.asd.tac.constellation.graph.schema.Schema;
+import au.gov.asd.tac.constellation.graph.schema.SchemaFactoryUtilities;
+import au.gov.asd.tac.constellation.graph.schema.visual.VisualSchemaFactory;
+import au.gov.asd.tac.constellation.graph.schema.visual.attribute.objects.Blaze;
+import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
+import au.gov.asd.tac.constellation.utilities.color.ConstellationColor;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.fail;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+/**
+ *
+ * @author antares
+ */
+public class BlazeContextMenuNGTest {
+    
+    private StoreGraph graph;
+    private int vxId1;
+    private int vxId2;
+    private int vxId3;
+    
+    private int blazeVertexAttribute;
+    private int selectedVertexAttribute;
+    
+    @BeforeClass
+    public static void setUpClass() throws Exception {
+    }
+
+    @AfterClass
+    public static void tearDownClass() throws Exception {
+    }
+
+    @BeforeMethod
+    public void setUpMethod() throws Exception {
+        final Schema schema = SchemaFactoryUtilities.getSchemaFactory(VisualSchemaFactory.VISUAL_SCHEMA_ID).createSchema();
+        graph = new StoreGraph(schema);
+        
+        vxId1 = graph.addVertex();
+        vxId2 = graph.addVertex();
+        vxId3 = graph.addVertex();
+        
+        blazeVertexAttribute = VisualConcept.VertexAttribute.BLAZE.ensure(graph);
+        selectedVertexAttribute = VisualConcept.VertexAttribute.SELECTED.ensure(graph);
+        
+        graph.setObjectValue(blazeVertexAttribute, vxId1, new Blaze(60, ConstellationColor.BANANA));
+        
+        graph.setBooleanValue(selectedVertexAttribute, vxId1, true);
+        graph.setBooleanValue(selectedVertexAttribute, vxId2, true);
+    }
+
+    @AfterMethod
+    public void tearDownMethod() throws Exception {
+    }
+
+    /**
+     * Test of getItems method, of class BlazeContextMenu. Node Element type
+     */
+    @Test
+    public void testGetItemsNodes() {
+        System.out.println("getItemsNodes");
+        
+        final List<ConstellationColor> colorList = Arrays.asList(ConstellationColor.BANANA, ConstellationColor.getColorValue(0.4F, 0.5F, 0.6F, 1F));
+        final List<String> expResult = Arrays.asList("Banana", "#668099", "Add Custom Blazes", "Remove Blazes");
+        
+        final BlazeContextMenu instance = new BlazeContextMenu();
+        try(final MockedStatic<BlazeActions> blazeActionsMockedStatic = Mockito.mockStatic(BlazeActions.class)) {
+            blazeActionsMockedStatic.when(() -> BlazeActions.getPresetCustomColors()).thenReturn(colorList);
+            final List<String> result = instance.getItems(null, GraphElementType.VERTEX, 0);
+            assertEquals(result, expResult);
+        }
+    }
+    
+    /**
+     * Test of getItems method, of class BlazeContextMenu. Transaction Element type
+     */
+    @Test
+    public void testGetItemsTransactions() {
+        System.out.println("getItemsTransactions");
+        
+        final BlazeContextMenu instance = new BlazeContextMenu();
+        final List<String> result = instance.getItems(null, GraphElementType.TRANSACTION, 0);
+        assertEquals(result, Collections.emptyList());
+    }
+}